+++ api/src/test/java/org/grnet/status/api/KeycloakTest.java
@@ -3,6 +3,8 @@ package org.grnet.status.api;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.keycloak.client.KeycloakTestClient;
 import io.restassured.RestAssured;
+import jakarta.inject.Inject;
+import org.grnet.status.services.TenantService;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.TestInstance;
 
@@ -14,12 +16,14 @@ import io.quarkus.test.common.http.TestHTTPResource;
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class KeycloakTest {
 
+    @Inject
+    TenantService tenantService;
     @TestHTTPResource
     URI baseUri;
-
     protected String adminToken;
     protected String aliceToken;
     protected String bobToken;
+
 
     KeycloakTestClient keycloakClient = new KeycloakTestClient();
 
@@ -29,6 +33,7 @@ public class KeycloakTest {
         adminToken = getAccessToken("admin");
         aliceToken = getAccessToken("alice");
         bobToken = getAccessToken("bob");
+        tenantService.deleteAll();
     }
 
     protected String getAccessToken(String username) {