+++ service/src/main/java/org/grnet/status/services/clients/ArgoWebApiClient.java
@@ -2,8 +2,13 @@ package org.grnet.status.services.clients;
 
 import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import org.grnet.status.dtos.argo.ArgoReportsResponse;
 import org.grnet.status.dtos.argo.ArgoStatusGroupsResponse;
+import org.grnet.status.dtos.tenant.TenantRequestDto;
+import org.grnet.status.dtos.tenant.TenantWebApiGetResponse;
+import org.grnet.status.dtos.tenant.TenantWebApiCreateResponse;
+import org.grnet.status.dtos.tenant.TenantWebApiUpdateResponse;
 
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
@@ -19,4 +24,26 @@ public interface ArgoWebApiClient {
             @HeaderParam("x-api-key") String apiKey,
             @PathParam("report") String report
     ) throws WebApplicationException, ProcessingException;
+
+    @POST
+    @Path("/api/v2/admin/tenants")
+    TenantWebApiCreateResponse createTenant(
+            @HeaderParam("x-api-key") String apiKey,
+            TenantRequestDto request
+    ) throws WebApplicationException, ProcessingException;
+
+    @GET
+    @Path("/api/v2/admin/tenants/{id}")
+    TenantWebApiGetResponse getTenant(
+            @HeaderParam("x-api-key") String apiKey,
+            String id
+    ) throws WebApplicationException, ProcessingException;
+
+    @PUT
+    @Path("/api/v2/admin/tenants/{id}")
+    TenantWebApiUpdateResponse updateTenant(String id,
+                                            @HeaderParam("x-api-key") String apiKey,
+                                            TenantRequestDto request
+    ) throws WebApplicationException, ProcessingException;
+
 }